First attempt at matching.
For this attempt, we are looking for matches between words in the model,
family, and manufacturer fields in products.txt and words in the title
field of listings.txt.  Case and accents are ignored.  Consecutive
combinations of words in listings.txt are allowed to match single words
in products.txt.  In order for a match to be reported, the algorithm
must find exactly one match, and must find at least one match for the
"model" field.