Second attempt at matching.
For this attempt, rather than attempting to resolve sets of product
matches as we are iterating through the words in the listing, we simply
note the nodes in the trie where matches exist, and then we resolve them
afterward.  The following rules are used to resolve the matches for a
trie:

1) A leaf node resolves to the set of products associated with it, if it
is a matching node.  If it is not, it resolves to the set of all
products (i.e., the "universe").
2) An internal node which has children that do NOT resolve to the
universe will resolve to the LARGER of:
   (a) The intersection of the resolving sets of size 1 of its children.
   (b) The intersection of the resolving sets of all of its children.
3) An internal node for which all of its children resolve to the
universe
will be treated as a leaf node.

To find a match, we first determine the resolving set for the root of
the manufacturer trie against the "manufacturer" string in the listing. 
We then find the resolving set of the model trie against the "title"
string in the listing.  If there is exactly one product in the resolving
set for the model trie, and if that product is also in the resolving set
for the manufacturer trie, then we report a match.  A more thorough
explanation for the reasons for these rules will be included in the
comments.

We also do some additional massaging of the data to begin with.  In
particular, we eliminate everything after the word "for" (or its french
translation, "pour"), as this word typically indicates that the product
is actually an accessory (battery, case, etc) *for* the camera, and not
the camera itself.