Changed the output format to conform to the specifications provided.
The results should be printed, one per line.  A single result is a json
object corresponding to a single product.  The result should have two
properties: "product_name", which is pulled from the corresponding
property in products.txt, and "listings", which is an array of the
matching listing json objects from listings.txt.