Removed a limitation on matching a certain class of products.
For the case where all strings which match one product also match another,
for example:

   (a) Pentax WG-1
   (b) Pentax WG-1 GPS

we could not ever match (a), because any string which would match (a) would
also match (b).  Now, when multiple results are obtained, the algorithm will
eliminate any matches which are not "maximal".  A match is maximal if there is
no string which may be appended to yield a longer match for the same product.
Thus, "WG-1" is a maximal match for (a) but not for (b).  So now, a listing
such as:

   "PENTAX Optio WG-1 14 MP Rugged Waterproof Digital Camera"

will yield a match against product (a), whereas before it would yield no
match because multiple results were obtained.